# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://price-hunter-backend.ddev.site'
paths:
...
  '/api/categories/{categoryName}':
    get:
      summary: 'Get products by category name.'
      operationId: getProductsByCategoryName
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 1
                  name: Laptop
                  category_id: 1
                  created_at: '2021-09-15T14:59:48.000000Z'
                  updated_at: '2021-09-15T14:59:48.000000Z'
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: Laptop
                  category_id:
                    type: integer
                    example: 1
                  created_at:
                    type: string
                    example: '2021-09-15T14:59:48.000000Z'
                  updated_at:
                    type: string
                    example: '2021-09-15T14:59:48.000000Z'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Category not found.'
                properties:
                  message:
                    type: string
                    example: 'Category not found.'
      tags:
        - Categories
      security: []
    parameters:
      -
        in: path
        name: categoryName
        description: 'The name of the category.'
        example: Electronics
        required: true
        schema:
          type: string
...